+++ src/main/java/io/appulse/utils/threads/FutureUtils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,6 +19,7 @@ package io.appulse.utils.threads;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
+import java.util.concurrent.Future;
 import java.util.function.Consumer;
 import java.util.stream.Stream;
 
@@ -68,6 +69,17 @@ public final class FutureUtils {
   public static <T> CompletableFuture<T> firstCompletedWithoutException (@NonNull List<? extends CompletionStage<? extends T>> futures) {
     val stream = futures.stream();
     return firstCompletedWithoutException(stream);
+  }
+
+  /**
+   * Transforms standard stdlib's {@link Future} instance into {@link CompletableFuture}.
+   *
+   * @param future for transform.
+   *
+   * @return {@link CompletableFuture} instance.
+   */
+  public static <T> CompletableFuture<T> toCompletableFuture (@NonNull Future<T> future) {
+    return new CompletablePromise<>(future);
   }
 
   // https://stackoverflow.com/a/34163913